+++ s4e-web/src/app/views/login/login.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {LoginComponent} from './login.component';
-import {ReactiveFormsModule} from '@angular/forms';
 import {ShareModule} from '../../common/share.module';
 import {FormErrorModule} from '../../components/form-error/form-error.module';