+++ s4e-web/src/app/views/settings/profile/change-password/change-password.component.ts
@@ -1,3 +1,5 @@
+import { ProfileStore } from './../../../../state/profile/profile.store';
+import { NotificationService } from './../../../../../../projects/notifications/src/lib/state/notification.service';
 import { Component } from '@angular/core';
 import { GenericFormComponent } from 'src/app/utils/miscellaneous/generic-form.component';
 import { ProfileQuery } from 'src/app/state/profile/profile.query';
@@ -6,7 +8,6 @@ import { AkitaNgFormsManager } from '@datorama/akita-ng-forms-manager';
 import { FormState } from 'src/app/state/form/form.model';
 import { Router } from '@angular/router';
 import { ProfileService } from 'src/app/state/profile/profile.service';
-import { SessionQuery } from 'src/app/state/session/session.query';
 import { FormGroup, Validators, FormControl } from '@ng-stack/forms';
 import { untilDestroyed } from 'ngx-take-until-destroy';
 
@@ -21,14 +22,16 @@ export class ChangePasswordComponent extends GenericFormComponent<ProfileQuery, 
     router: Router,
     private _profileService: ProfileService,
     private _profileQuery: ProfileQuery,
+    private _profileStore: ProfileStore,
+    private _notificationService: NotificationService
   ) {
     super(fm, router, _profileQuery, 'resetPassword');
   }
 
   ngOnInit(): void {
     this.form = new FormGroup<PasswordChangeFormState>({
-      oldPassword: new FormControl<string>(null, [Validators.required]),
-      newPassword: new FormControl<string>(null, [Validators.required])
+      oldPassword: new FormControl<string>(null, [Validators.required, Validators.minLength(6)]),
+      newPassword: new FormControl<string>(null, [Validators.required, Validators.minLength(6)])
     });
     super.ngOnInit();
   }
@@ -41,7 +44,18 @@ export class ChangePasswordComponent extends GenericFormComponent<ProfileQuery, 
           this.form.controls.newPassword.value
         )
         .pipe(untilDestroyed(this))
-        .subscribe();
+        .subscribe(() => {
+          this._notificationService.addGeneral({
+            type: 'success',
+            content: 'Hasło zostało zmienione'
+          });
+          this.reset();
+        });
     }
+  }
+
+  reset() {
+    this.form.reset();
+    this._profileStore.setError(null);
   }
 }