+++ s4e-web/src/app/views/settings/profile/profile.module.ts
@@ -4,6 +4,7 @@ import {ShareModule} from '../../../common/share.module';
 import {ProfileComponent} from './profile.component';
 import { ChangePasswordComponent } from './change-password/change-password.component';
 import { S4EFormsModule } from 'src/app/form/form.module';
+import { FormErrorModule } from 'src/app/components/form-error/form-error.module';
 
 @NgModule({
   declarations: [
@@ -12,7 +13,8 @@ import { S4EFormsModule } from 'src/app/form/form.module';
   ],
   imports: [
     ShareModule,
-    S4EFormsModule
+    S4EFormsModule,
+    FormErrorModule
   ],
   exports: [
     ProfileComponent,