+++ CHANGELOG.md
@@ -41,6 +41,8 @@ The most recent changes are on top, in each type of changes category.
 
 ### Fixed
 
+- Fix Institution edition, registration form, add child institution and cancel password change [#558](https://github.com/cyfronet-fid/sat4envi/pull/558)
+- [fix] dropdowns toggle on map and in settings [#562](https://github.com/cyfronet-fid/sat4envi/issues/562)
 - Fix institution loading in search box
 - Fix registration form payload [#537](https://github.com/cyfronet-fid/sat4envi/pull/538)
 - Handle nulls in custom Validators [#491](https://github.com/cyfronet-fid/sat4envi/issues/491)